% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTest.R
\name{adda}
\alias{adda}
\title{Add together two numbers.}
\usage{
adda(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Add together two numbers.
}
\examples{
adda(1, 1)
adda(10, 1)
}
